pyproject.toml
[project]
name = "aai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
]

[tool.pytest.ini_options]
pythonpath=["."]
testpaths=["tests"]

[dependency-groups]
dev = [
    "polyfactory>=2.21.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
]
